feat(tests): cache file.
