+++ tests_integration/test_intro.py
@@ -1,7 +1,23 @@
+import os
+from pathlib import Path
 from typing import Any
 
+import pytest
+
 import dspy
-from tests_integration.base import BaseIntegrationTestWithCache
+
+
+class BaseIntegrationTestWithCache:
+    @pytest.fixture(autouse=True)
+    def setup(self) -> None:
+        # Base directory for all DSPy modules
+        library_dir = Path(__file__).resolve().parent
+        base_dir = library_dir.parent
+        cache_dir = str(base_dir / "cache")
+        os.environ["DSP_NOTEBOOK_CACHEDIR"] = cache_dir
+
+        if cache_dir and not Path(cache_dir).exists():
+            Path(cache_dir).mkdir(parents=True)
 
 
 class TestIntroIntegration(BaseIntegrationTestWithCache):
